backend/requirements.txt
flask
tensorflow
numpy
flask_cors
Pillow